Bug 1643557 - Make pk11_FindObjectByTemplate accept a size_t count rather than a signed type to avoid internal signed-unsigned comparison warnings.  r=kjacobs
Depends on D78454

Differential Revision: https://phabricator.services.mozilla.com/D78455

--HG--
extra : moz-landing-system : lando

+++ lib/pk11wrap/pk11skey.c
@@ -570,7 +570,7 @@ PK11_FindFixedKey(PK11SlotInfo *slot, CK_MECHANISM_TYPE type, SECItem *keyID,
     CK_ATTRIBUTE *attrs;
     CK_BBOOL ckTrue = CK_TRUE;
     CK_OBJECT_CLASS keyclass = CKO_SECRET_KEY;
-    int tsize = 0;
+    size_t tsize = 0;
     CK_OBJECT_HANDLE key_id;
 
     attrs = findTemp;
