+++ lib/pk11wrap/secmodi.h
@@ -93,7 +93,7 @@ CK_RV pk11_notify(CK_SESSION_HANDLE session, CK_NOTIFICATION event,
                   CK_VOID_PTR pdata);
 void pk11_SignedToUnsigned(CK_ATTRIBUTE *attrib);
 CK_OBJECT_HANDLE pk11_FindObjectByTemplate(PK11SlotInfo *slot,
-                                           CK_ATTRIBUTE *inTemplate, int tsize);
+                                           CK_ATTRIBUTE *inTemplate, size_t tsize);
 CK_OBJECT_HANDLE *pk11_FindObjectsByTemplate(PK11SlotInfo *slot,
                                              CK_ATTRIBUTE *inTemplate, size_t tsize, int *objCount);
